requirements.txt
altair==5.0.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
cachetools==5.3.1
catalogue==1.0.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
cryptography==40.0.2
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
docx2txt==0.8
en-core-web-sm==2.3.1
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
fire==0.5.0
Flask==2.3.2
fonttools==4.39.3
fqdn==1.5.1
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
murmurhash==1.0.9
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pdf2docx==0.5.6
pdfminer.six==20221105
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
plac==1.1.3
platformdirs==3.5.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.17
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyMuPDF==1.22.2
pyresparser==1.0.6
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2023.5.4
requests==2.30.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==2.3.5
srsly==1.0.6
stack-data==0.6.2
streamlit==1.23.1
tenacity==8.2.2
termcolor==2.3.0
terminado==0.17.1
thinc==7.4.6
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
urllib3==1.26.15
validators==0.20.0
wasabi==0.10.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.3
widgetsnbextension==4.0.7
wsproto==1.2.0
yarg==0.1.9
zipp==3.15.0